Introduce architecture differentiation
In order to properly handle building for different architectures, a
value for --host for configure, architecture name, as well as a value
for baseline toolchain image is injected into the dockerfiles as args.
This way the same Dockerfile can be used for multiple arch targets.

This also results in a new naming convetion for the containers, where
before they were *name:version*, now they're *name:version-arch*.
Makefile rules names also changes, from *name/version* to
*name-arch:version*. This is backwards from the container names, because
of the limiting ways to manipulate strings in makefiles. Keeping just
the version as a variability in the rules, leads to them being simpler,
while still being flexible enough to cover each version.