Updates rule for OpenSSL
Following the existing arch differentiations, the openssl rule is split
into an armv7-a and an x86 variant.

Updates and cleanup of the openssl dockerfile has been made accordingly.
Since the build configuration of openssl is a lot more involved, than
some of the other rules', it's worth noting the number of extra build
variables added, in order to get a working rule for both architectures.