Updates sqlite3
This adds a patch for Makefile.am to disabled versioned builds of the
library through the -avoid-version option to libtool.

Previously, when libspatialite would build against the versioned variant
of sqlite3, it would link against a `libsqlite3.so.0` version of the
library, which does not conform to lib loading on android and would thus
fail at run time.