Streamlines builds and tests
This change get's rid of the all-in-one STAND_TAG build arg, in favour
of injecting PLATFORM and TOOLCHAIN parts of stand image tags
separately. This is to easier fascilitate x86 builds.

All library builds involved in this change has also been tested for x86
compatibility, where in some cases a few changes to the test programs
have been done as well.